Refactor data-utils to separate api from factory
