+++ playwright-e2e/data-utils/api/consented/ConsentedEventApi.ts
@@ -1,6 +1,6 @@
-import { ccdApi } from "../../fixtures/fixtures";
-import config from "../../config/config";
-import { ConsentedEvents, CaseType, PayloadPath } from "../../config/case-data";
+import { ccdApi } from "../../../fixtures/fixtures";
+import config from "../../../config/config";
+import { ConsentedEvents, CaseType, PayloadPath } from "../../../config/case-data";
 
 export class ConsentedEventApi {
   private static async updateCaseWorkerSteps(
