+++ playwright-e2e/fixtures/fixtures.ts
@@ -1,7 +1,7 @@
 import { mergeTests, mergeExpects } from '@playwright/test';
 import { test as create } from './create-fixture';
 import { test as a11yTest, expect as a11yExpect } from './axe-fixture';
-import { CcdApi } from '../data-utils/CcdApi';
+import { CcdApi } from '../data-utils/api/CcdApi';
 
 export const test = mergeTests(create, a11yTest);
 export const expect = mergeExpects(a11yExpect);
