+++ playwright-e2e/test/consented/application_payment_submission.spec.ts
@@ -2,7 +2,7 @@ import { test } from '../../fixtures/fixtures'
 import config from '../../config/config';
 import { ConsentedEvents } from '../../config/case-data';
 import { paymentDetailsTabData } from '../../resources/tab_content/payment_details_tabs';
-import { ConsentedCaseFactory } from '../../data-utils/consented/ConsentedCaseFactory';
+import { ConsentedCaseFactory } from '../../data-utils/factory/consented/ConsentedCaseFactory';
 
 test(
     'Consented - Application Payment Submission',
