+++ playwright-e2e/test/consented/hwf_application_accepted.spec.ts
@@ -1,7 +1,7 @@
 import { test } from '../../fixtures/fixtures';
 import config from '../../config/config';
 import { ConsentedEvents } from '../../config/case-data';
-import { ConsentedCaseFactory } from '../../data-utils/consented/ConsentedCaseFactory';
+import { ConsentedCaseFactory } from '../../data-utils/factory/consented/ConsentedCaseFactory';
 
 
 test(
