+++ playwright-e2e/test/contested/FR_addSchedulingListingInfo/list_for_hearing_express_case.spec.ts
@@ -1,6 +1,6 @@
 import { expect, test } from '../../../fixtures/fixtures';
 import config from '../../../config/config';
-import { ContestedCaseFactory } from '../../../data-utils/contested/ContestedCaseFactory';
+import { ContestedCaseFactory } from '../../../data-utils/factory/contested/ContestedCaseFactory';
 import { ContestedEvents } from '../../../config/case-data';
 
 async function performListForHearingFlow(
