+++ playwright-e2e/test/contested/FR_amendApplication/amend_paper_application_into_express_pilot.spec.ts
@@ -1,6 +1,6 @@
 import { test } from '../../../fixtures/fixtures';
 import config from '../../../config/config';
-import { ContestedCaseFactory } from '../../../data-utils/contested/ContestedCaseFactory';
+import { ContestedCaseFactory } from '../../../data-utils/factory/contested/ContestedCaseFactory';
 import { ContestedEvents } from '../../../config/case-data';
 
 test.describe('Contested - Paper Case - Amend application into Express Pilot', () => {
