+++ playwright-e2e/test/contested/FR_listForInterimHearing/list_for_interim.hearing.spec.ts
@@ -2,7 +2,7 @@ import { expect, test } from '../../../fixtures/fixtures';
 import config from '../../../config/config';
 import { ContestedEvents } from '../../../config/case-data';
 import { YesNoRadioEnum } from '../../../pages/helpers/enums/RadioEnums';
-import { ContestedCaseFactory } from '../../../data-utils/contested/ContestedCaseFactory';
+import { ContestedCaseFactory } from '../../../data-utils/factory/contested/ContestedCaseFactory';
 
 // Hearing types
 const MaintenancePendingSuit = {
