+++ playwright-e2e/test/contested/FR_manageExpressCase/manage_express_case.spec.ts
@@ -1,7 +1,7 @@
 import { test } from '../../../fixtures/fixtures';
 import config from '../../../config/config';
 import { ContestedEvents } from '../../../config/case-data';
-import { ContestedCaseFactory } from '../../../data-utils/contested/ContestedCaseFactory';
+import { ContestedCaseFactory } from '../../../data-utils/factory/contested/ContestedCaseFactory';
 
 test.describe('Contested - Manage Express Case', () => {
   test(
