+++ playwright-e2e/test/contested/assign_respondent_case.spec.ts
@@ -1,8 +1,8 @@
 import { test } from '../../fixtures/fixtures';
 import config from '../../config/config';
 import { respondentAssignedCaseTabs } from '../../resources/tab_content/contested/respondent_assigned_case_tabs';
-import { assignCaseToRespondent } from '../../data-utils/CaseAssignmentHelper';
-import { ContestedCaseFactory } from '../../data-utils/contested/ContestedCaseFactory';
+import { assignCaseToRespondent } from '../../pages/helpers/CaseAssignmentHelper';
+import { ContestedCaseFactory } from '../../data-utils/factory/contested/ContestedCaseFactory';
 
 test(
   'Contested - Respondent Assigned Case',
